Add files via upload
A half adder is a combinational circuit that adds two single-bit binary numbers and produces a sum and carry as output. It has two inputs and two outputs, one for the sum (S) and one for the carry (C). A half adder can be implemented using basic logic gates such as AND, XOR, and NOT gates. 
A UVM monitor is a passive component used to capture DUT signals using a virtual interface and translate them into a sequence item format. 
A reference model is also called a predictor and would mimic the functionality of the design. After receiving data objects, it can either perform calculations and predict the expected value or send it to a reference model to get expected values